requirements.txt
llama_index_core
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
llama-index-llms-groq
llama-index-retrievers-bm25
llama-index-readers-llama-parse
streamlit
pandas
deepeval